examples/production/04-cli-tool-generator/requirements.txt
# CLI Tool Generator Dependencies
# Install with: pip install -r requirements.txt

# Core CLI framework
click>=8.1.0

# JSON Schema validation
jsonschema>=4.20.0

# YAML parsing for specifications
pyyaml>=6.0.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking (optional but recommended)
mypy>=1.7.0
types-PyYAML>=6.0.0
